Add files via upload
The data file was modified, and an arrow tile was added at the end. The code is expanded, and I have removed many of the now not used code. 
I have started to write the new menu control code. Said arrow tile is now a sprite, and it shows on the options screen. 
A new titlescreen has been added as well, along with the decompression routines, and a few helpers, since the ones the game uses has some things that are not directly compatible with how I wanted to use them.